KE: Fixed routes for new pages.
