+++ client/src/App.css
@@ -5,10 +5,16 @@
   flex-direction: column;
   align-items: center;
   height: 100vh;
-  overflow-y: auto;
+  overflow-y: hidden;
+  overflow-x: hidden;
   background-color: #002;
   color: white;
   font-family: 'Roboto', Verdana, Geneva, Tahoma, sans-serif;
+}
+
+.appBody {
+  overflow-y: auto;
+  padding: 1vw 10vw;
 }
 
 h1 {
@@ -75,7 +81,7 @@ button:active {
   color: white;
 }
 
-/* Header */
+/* Header and Footer */
 
 .header, .footer {
   width: 100vw;
@@ -83,13 +89,13 @@ button:active {
   flex-direction: column;
   align-items: center;
   justify-content: center;
-  margin: 0.5vw 0;
 }
 
 .nav {
-  width: 40vw;
+  width: 50vw;
   display: flex;
-  justify-content: space-around;
+  align-items: center;
+  justify-content: space-evenly;
 }
 
 /* Searchbar */
