--- client/src/components/PlacesToGo.jsx
@@ -1,37 +1,0 @@
-const PlacesToGo = () => {
-    return (   
-    <div>
-    <h1>While You're Here...</h1>
-    <h3>Check out these suggested local venues!</h3>
-    <h4>Current promotional offers:</h4>
-    <p>Pizza Metro: Enjoy a Buy-One-Get-One-Free offer on all vegan pizzas by presenting a QA Cinemas ticket dated between Aug 2041 and Nov 2042.</p>
-    <p>Dominus Chicken: 50% off on all non-chicken meals when ordered by a ticket-holder for "FAST AND FURIOUS 25: THE LAST RIDE". Ends 24/09/2042.</p>
-
-    <h4><b>Pizza Metro</b></h4>
-    <img alt="Interior of Pizza Metro" src="https://c8.alamy.com/comp/2HA30YB/interior-chicago-style-deep-dish-pizza-restaurant-japes-soho-london-uk-2HA30YB.jpg" width={600} height={400} />
-    <h5>Address: 132 Nearfar Avenue, Lower Midley, RP40 Z42</h5>
-    <h5>Contact: 01255-592384</h5>
-    <p>Pizza Metro is a local pizzeria. Founded in 2028, their formulaic 3D-printed 
-        dough technology will leave you feeling adequately satisfied every time.</p>
-
-    <h4><b>Dominus Chicken</b></h4>
-    <img alt="Interior of Dominus Chicken" src="https://thumbs.dreamstime.com/b/interior-fast-food-take-out-restaurant-mall-johannesburg-south-africa-july-137197867.jpg" width={600} height={400} />
-    <h5>Address: 12a Last Road, The City, RC21 X30</h5>
-    <h5>Contact: 15825-582582</h5>
-    <p>If chicken's your thing, the Dominus has you covered. Featuring free-range chicken
-        freshly slaughtered in-store, each bite contains freshly-killed flavour. That's the Dominus Promise (trademark pending).</p> 
-
-    <h4>McKing</h4> 
-    <img alt="Logo of McKing" src="https://i.ytimg.com/vi/fhBJ_hGMo8U/hqdefault.jpg" width={600} height={400} />
-    <h5>Address: 34 Quisitions Lane, Upper Midley, RP35 G59</h5>
-    <h5>Contact: 85772-583762</h5>
-    <p>Fast-food lovers were ecstatic when historic fast-food chain McDonalds announced 
-        its hostile acquisition of Burger King in 2040. Now, the two chains have joined 
-        forces to bring you excellent flavour from all kinds of raw materials. Just like 
-        their new motto says, we're lovin' the King!</p>
-    </div>
-    
-    )
-};
-
-export default PlacesToGo;
