+++ client/src/components/structure/Header.jsx
@@ -10,10 +10,10 @@ const Header = () => {
             <div className="nav">
               <button type="button" onClick={() => navigator('/')}>Home</button>
               <button type="button" onClick={() => navigator('/classifications')}>Movie Classifications</button>
-              <button type="button" onClick={() => navigator('/placesToVisit')}>Places To Visit</button>
+              <button type="button" onClick={() => navigator('/placesToGo')}>Places To Go</button>
               <button><s>Discussion Board</s></button>
-              <button><s>Contact Us</s></button>
-              <button><s>ETC, ETC, ETC</s></button>
+              <button type="button" onClick={() => navigator('/contact')}>Contact Us</button>
+              <button type="button" onClick={() => navigator('/openingTimes')}>Opening Times</button>
             </div>
         </div>
     );
